+++ src/app/pages/recursos-humanos/conciliacion-rh/conciliacion-rh.component.html
@@ -11,7 +11,7 @@
           <div class="form-group row">
             <label class="col-md-2 col-form-label" for="file-input">Subir Archivo</label>
             <div class="col-md-9">
-            <input id="file-input" type="file" accept=".xlsx, .xls, .csv  " name="file-input" (change)="onFileChange($event.target.files)" class="file-input">
+            <input #fileInput id="file-input" type="file" accept=".xlsx, .xls, .csv  " name="file-input" (change)="onFileChange($event.target.files)" class="file-input">
             </div>
             </div>
 
@@ -48,8 +48,10 @@
           <div class="table-responsive slide_1" *ngIf="conciliacionProcesados.length > 0" >
             
             <!-- Tabla correctos-->
+            
+            <h3 *ngIf="conciliacionCorrectos.length > 0"> Ahorradores Validados</h3>
             <table class="table table-bordered table-sm" *ngIf="conciliacionCorrectos.length > 0">
-              <caption>Ahorradores Correctos</caption>
+              
               <thead class="thead-dark">
                 <tr>
                   <th scope="col" class="text-center">Clave</th>
@@ -62,15 +64,14 @@
               <tbody>
                     <tr *ngFor = "let fila of conciliacionCorrectos" >
                       <td scope="row" class="text-left"> {{fila.noEmpleado}} </td>
-                      <td scope="row" class="text-left"> {{fila.nombre}} </td>
+                      <td scope="row" class="text-left"> {{fila.nombre | uppercase }} </td>
                       <td scope="row" class="text-left"> {{fila.saldo}} </td>
-                      <td scope="row" class="text-left"> {{fila.observaciones}} </td>
+                      <td scope="row" class="text-left"> Validado </td>
                     </tr>
               </tbody>
-            </table> 
-
-            
-            <p *ngIf="conciliacionErroneos.length > 0"> Ahorradores Erroneos</p>
+            </table>
+            <!-- Tabla Erroneos-->
+            <h3 *ngIf="conciliacionErroneos.length > 0"> Ahorradores No Validados</h3>
             <table class="table table-bordered table-sm" *ngIf="conciliacionErroneos.length > 0">
               
               <thead class="thead-dark">
@@ -87,7 +88,7 @@
                       <td scope="row" class="text-left"> {{fila.noEmpleado}} </td>
                       <td scope="row" class="text-left"> {{ fila.nombre | uppercase }} </td>
                       <td scope="row" class="text-left"> {{fila.saldo}} </td>
-                      <td scope="row" class="text-left"> {{fila.observaciones}} </td>
+                      <td scope="row" class="text-left"> {{fila.observaciones | uppercase }} </td>
                     </tr>
               </tbody>
             </table>
